+++ tests/unit/source_code/test_whitelist.py
@@ -1,4 +1,4 @@
-from databricks.labs.ucx.source_code.dependencies import SourceContainer
+from databricks.labs.ucx.source_code.dependency_loaders import SourceContainer
 from databricks.labs.ucx.source_code.whitelist import Whitelist, UCCompatibility
 from tests.unit import _load_sources
